Add files via upload
This is the full project with all the functionalities I have made up to this date. It can send emails, but I have an error causing it to display html in the email itself that I just can't seem to figure out how to fix.

The first 10 files are to be put into a mysite folder in the base directory.

The rest must be in a folder titled Glizzymail. The Glizzymail _pychache_, migrations, and .html files must all be in their own folders.
Name the .html file folder as templates.

This version requires at least python 3.8 and Django installed in the base directory of the first folder for the website.